requirements.txt
anthropic==0.32.0
google_api_python_client==2.139.0
python-dotenv==1.0.1
pytube==15.0.0
youtube_transcript_api==0.6.2
